lecture_notes_generator/requirements.txt
pdfplumber==0.11.0
python-pptx==0.6.23
sentence-transformers==2.7.0
faiss-cpu==1.8.0
openai==1.25.0
pydantic==2.7.1
